backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
typhoon-ocr
python-dotenv>=1.0.1
pypdf>=4.0.0
